Move networking sources into the clp namespace.
